backend/requirements.txt
django
djangorestframework
web3
py-solc-x